Additional tests

+++ steg/steg_decode_test.go
@@ -75,47 +75,46 @@ func TestMultiCarrierDecodeOrderMatters(t *testing.T) {
 }
 
 func TestDecodeByFileNames(t *testing.T) {
-	err := steg.DecodeByFileNames("../examples/test_decode.jpeg", "result")
-	if err != nil {
-		t.Fatalf("Error decoding file: %v", err)
-	}
-
-	defer func() {
-		err = os.Remove("result")
-		if err != nil {
-			t.Fatalf("Error removing result file: %v", err)
+	AssertDecodeByFileNames(t, []string{"../examples/test_decode.jpeg"}, "../examples/lake.jpeg", func(strings []string, s string) {
+		if len(strings) != 1 {
+			t.Fatalf("Exactly one carrier expected")
 		}
-	}()
+		err := steg.DecodeByFileNames(strings[0], s)
+		if err != nil {
+			t.Fatalf("Error decoding file: %v", err)
+		}
+	})
+}
 
-	wanted, err := os.Open("../examples/lake.jpeg")
-	if err != nil {
-		t.Fatalf("Error opening file examples/lake.jpg: %v", err)
-	}
-	defer wanted.Close()
-
-	result, err := os.Open("result")
-	if err != nil {
-		t.Fatalf("Error opening file result: %v", err)
-	}
-	defer result.Close()
-
-	wantedBytes, err := ioutil.ReadAll(wanted)
-	if err != nil {
-		t.Fatalf("Error reading file examples/lake.jpg: %v", err)
-	}
-
-	resultBytes, err := ioutil.ReadAll(result)
-	if err != nil {
-		t.Fatalf("Error reading file result: %v", err)
-	}
-
-	if !bytes.Equal(wantedBytes, resultBytes) {
-		t.Error("Assertion failed!")
-	}
+func TestMultiCarrierDecodeByFileNames(t *testing.T) {
+	AssertDecodeByFileNames(t, []string{"../examples/test_multi_carrier_decode1.jpeg", "../examples/test_multi_carrier_decode2.jpeg"}, "../examples/video.mp4", func(strings []string, s string) {
+		err := steg.MultiCarrierDecodeByFileNames(strings, s)
+		if err != nil {
+			t.Fatalf("Error decoding file: %v", err)
+		}
+	})
 }
 
 func TestDecodeByFileNamesShouldReturnErrorWhenCarrierFileMissing(t *testing.T) {
 	err := steg.DecodeByFileNames("not_existing_file", "result")
+	if err == nil {
+		os.Remove("result")
+		t.FailNow()
+	}
+	t.Log(err)
+}
+
+func TestMultiCarrierDecodeByFileNamesShouldReturnErrorWhenCarrierFileMissing(t *testing.T) {
+	err := steg.MultiCarrierDecodeByFileNames([]string{"not_existing_file1", "non_existing_file2"}, "result")
+	if err == nil {
+		os.Remove("result")
+		t.FailNow()
+	}
+	t.Log(err)
+}
+
+func TestMultiCarrierDecodeByFileNamesShouldReturnErrorWhenNoCarrierFilesProvided(t *testing.T) {
+	err := steg.MultiCarrierDecodeByFileNames([]string{}, "result")
 	if err == nil {
 		os.Remove("result")
 		t.FailNow()
@@ -133,6 +132,30 @@ func TestDecodeShouldReturnErrorWhenCarrierFileIsNotImage(t *testing.T) {
 	var result bytes.Buffer
 	err = steg.Decode(carrier, &result)
 	if err == nil {
+		t.FailNow()
+	}
+	t.Log(err)
+}
+
+func TestMultiCarrierDecodeShouldReturnErrorWhenCarrierFileIsNotImage(t *testing.T) {
+	carrier, err := os.Open("../README.md")
+	if err != nil {
+		t.Fatalf("Error opening carrier file: %v", err)
+	}
+	defer carrier.Close()
+
+	var result bytes.Buffer
+	err = steg.MultiCarrierDecode([]io.Reader{carrier}, &result)
+	if err == nil {
+		t.FailNow()
+	}
+	t.Log(err)
+}
+
+func TestMultiCarrierDecodeByFileNamesShouldReturnErrorWhenCarrierFileIsNotImage(t *testing.T) {
+	err := steg.MultiCarrierDecodeByFileNames([]string{"../README.md"}, "result")
+	if err == nil {
+		os.Remove("result")
 		t.FailNow()
 	}
 	t.Log(err)
@@ -179,6 +202,38 @@ func AssertDecode(t *testing.T, carrierNames []string, expected string, f func([
 	}
 
 	if bytes.Equal(wantedBytes, resultBytes) != shouldEqual {
+		t.Error("Assertion failed!")
+	}
+}
+
+func AssertDecodeByFileNames(t *testing.T, carrierNames []string, expected string, f func([]string, string)) {
+	f(carrierNames, "result")
+
+	defer os.Remove("result")
+
+	wanted, err := os.Open(expected)
+	if err != nil {
+		t.Fatalf("Error opening file examples/lake.jpg: %v", err)
+	}
+	defer wanted.Close()
+
+	result, err := os.Open("result")
+	if err != nil {
+		t.Fatalf("Error opening file result: %v", err)
+	}
+	defer result.Close()
+
+	wantedBytes, err := ioutil.ReadAll(wanted)
+	if err != nil {
+		t.Fatalf("Error reading file examples/lake.jpg: %v", err)
+	}
+
+	resultBytes, err := ioutil.ReadAll(result)
+	if err != nil {
+		t.Fatalf("Error reading file result: %v", err)
+	}
+
+	if !bytes.Equal(wantedBytes, resultBytes) {
 		t.Error("Assertion failed!")
 	}
 }
